requirements.txt
# Core dependencies for VLA Frequency Attack

# PyTorch (install with CUDA support as needed)
torch>=2.0.0
torchvision>=0.15.0

# Transformers and HuggingFace
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.8.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# TensorFlow (for preprocessing - matching OpenVLA)
tensorflow>=2.13.0

# LIBERO dependencies
robosuite>=1.4.0
mujoco>=2.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0

# OpenVLA specific
timm>=0.9.0
einops>=0.7.0
